move calculations into commodity_row_generator
