+++ test/fixtures/fixed_assets/commodities.yml
@@ -6,7 +6,7 @@ commodity_one:
   modified_by: joe
   amount: 10000
   status: A
-  activated_at: <%= Date.today %>
+  activated_at: <%= Date.today.change(month: 7, day: 1) %>
   planned_depreciation_type: P
   planned_depreciation_amount: 35
   btl_depreciation_type: B
