+++ test/models/fixed_assets/commodity_test.rb
@@ -109,86 +109,6 @@ class FixedAssets::CommodityTest < ActiveSupport::TestCase
     assert_equal 1001.0, @commodity.deprecated_evl_amount
   end
 
-  test 'should calculate with divide_to_payments' do
-    # We have a 12 month fiscal year
-    fiscal_year = @commodity.company.months_in_current_fiscal_year
-    assert_equal fiscal_year, 12
-
-    # We active in july
-    @commodity.activated_at = DateTime.now.change(month: 7, day: 1)
-
-    # Divide 1000 into 12 depreciations
-    result = @commodity.divide_to_payments(1000, 12)
-
-    # We should get 500 this fiscal year, in 6 depreciations
-    assert_equal 500, result.sum
-    assert_equal 6, result.count
-    assert_equal 83.33.to_d, result.first
-    assert_equal 83.35.to_d, result.last
-  end
-
-  test 'should calculate with fixed by percentage' do
-    # Tasapoisto vuosiprosentti
-    full_amount = 10000
-    percentage = 35
-
-    # We active in july
-    @commodity.activated_at = DateTime.now.change(month: 7, day: 1)
-
-    result = @commodity.fixed_by_percentage(full_amount, percentage)
-
-    assert_equal result.sum, full_amount * percentage / 100
-    assert_equal result.first, 833.33
-    assert_equal result.second, 833.33
-    assert_equal result.last, 166.68
-  end
-
-  test 'should calculate with degressive by percentage' do
-    # Menojäännöspoisto vuosiprosentti
-    # Full amount to be reducted
-    reduct = 10000
-
-    # We have a 12 month fiscal year
-    fiscal_year = @commodity.company.months_in_current_fiscal_year
-    assert_equal fiscal_year, 12
-
-    # We active in july
-    @commodity.activated_at = DateTime.now.change(month: 7, day: 1)
-
-    # Yearly degressive percentage
-    fiscalyearly_percentage = 35
-
-    result = @commodity.degressive_by_percentage(reduct, fiscalyearly_percentage)
-    assert_equal 6, result.count
-    assert_equal 3500, result.sum
-
-    result = @commodity.degressive_by_percentage(reduct, fiscalyearly_percentage, 3500)
-    assert_equal 6, result.count
-    assert_equal 2275, result.sum
-  end
-
-  test 'should calculate with fixed by month' do
-    # Tasapoisto kuukausittain
-    # Full amount to be reducted
-    total_amount = 10000
-    # Total amounts of depreciations
-    total_depreciations = 12
-
-    # We have a 12 month fiscal year
-    fiscal_year = @commodity.company.months_in_current_fiscal_year
-    assert_equal fiscal_year, 12
-
-    # We active in july
-    @commodity.activated_at = DateTime.now.change(month: 7, day: 1)
-
-    fiscal_year = @commodity.company.months_in_current_fiscal_year
-
-    result = @commodity.fixed_by_month(total_amount, total_depreciations, 6, 5000)
-
-    assert_equal 6, result.count
-    assert_equal 5000, result.sum
-  end
-
   test 'should calculate SUMU depreciation with fixed by percentage' do
     params = {
       voucher: nil,
